Update check compilation
