+++ checkCompilation.py
@@ -23,17 +23,19 @@ def parse_args():
     parser.add_argument("--syntax-only", dest="syntax_only", action="store_true", default=False)
     parser.add_argument("-j", dest="j", action="store", type=int, default=multiprocessing.cpu_count())
     parser.add_argument("--tests-file", dest="tests", action="store", type=str, default="")
+    parser.add_argument("--even-slow-models", dest="slow_models", action="store_true")
     return parser.parse_args()
 
 if __name__ == "__main__":
     args = parse_args()
 
-    models = None
-
     if args.tests == "":
         models = find_files("*.stan", args.directories)
     else:
         models, _ = read_tests(args.tests)
+
+    if not args.slow_models:
+        models = filter_out_weekly_models(models)
 
     executables = [m[:-5] for m in models]
     delete_temporary_exe_files(executables)
